Added unitless type, which is just an implementation of Quantity for f64
