+++ src/types.rs
@@ -4,6 +4,7 @@
 pub use self::pressures::*;
 pub use self::specific_energy::*;
 pub use self::temperatures::*;
+pub use self::unitless::*;
 pub use self::winds::*;
 
 /// A quantity is a common super trait for types that represent units of measurement.
@@ -96,6 +97,7 @@ macro_rules! implOpsForQuantity {
 mod pressures;
 mod specific_energy;
 mod temperatures;
+mod unitless;
 mod winds;
 
 #[cfg(all(test, feature = "use_optional"))]
